refactor(ui): split task pane edit form code
